Day 23 in progress

+++ src/main/scala/challenge/Day21.scala
@@ -11,7 +11,9 @@ object Day21 extends Challenge {
     // reasoned through trial and error, maybe at some point make some heuristic for automatically generating
     // working input
     val input =
-      List("NOT D T", "OR C T", "AND A T", "NOT T J", "WALK").flatMap(_.toList :+ '\n').map(_.toInt)
+      List("NOT D T", "OR C T", "AND A T", "NOT T J", "WALK")
+        .flatMap(_.toList :+ '\n')
+        .map(_.toLong)
 
     val output = ic.execute(ic.Input(program, input))
     // Console.err.println(output.sig.map(_.toChar).mkString)
